requirements.txt
Flask
Werkzeug
requests
ffmpeg-python
openai-whisper
gunicorn
APScheduler
srt
numpy
torch
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
psutil
boto3
Pillow
matplotlib
minio==7.1.0
python-dotenv==1.0.0